Avoid exposing new property methods in EclipseJdt
