+++ platforms/ide/ide/src/main/java/org/gradle/plugins/ide/eclipse/model/EclipseJdt.java
@@ -19,9 +19,7 @@ package org.gradle.plugins.ide.eclipse.model;
 import groovy.lang.Closure;
 import groovy.lang.DelegatesTo;
 import org.gradle.api.Action;
-import org.gradle.api.Incubating;
 import org.gradle.api.JavaVersion;
-import org.gradle.api.provider.Property;
 import org.gradle.plugins.ide.api.PropertiesFileContentMerger;
 import org.gradle.util.internal.ConfigureUtil;
 
@@ -78,9 +76,7 @@ public abstract class EclipseJdt {
      * <p>
      * For example see docs for {@link EclipseJdt}
      */
-    public JavaVersion getSourceCompatibility() {
-        return getSourceCompatibilityProperty().get();
-    }
+    public abstract JavaVersion getSourceCompatibility();
 
     /**
      * Sets source compatibility.
@@ -91,41 +87,14 @@ public abstract class EclipseJdt {
         setSourceCompatibility((Object) sourceCompatibility);
     }
 
-    public void setSourceCompatibility(Object sourceCompatibility) {
-        JavaVersion version = JavaVersion.toVersion(sourceCompatibility);
-        if (version != null) {
-            this.getSourceCompatibilityProperty().set(version);
-        }
-    }
-
-    /**
-     * The source Java language level.
-     * <p>
-     * For example see docs for {@link EclipseJdt}
-     *
-     * @since 9.0
-     */
-    @Incubating
-    abstract public Property<JavaVersion> getSourceCompatibilityProperty();
+    abstract public void setSourceCompatibility(Object sourceCompatibility);
 
     /**
      * The target JVM to generate {@code .class} files for.
      * <p>
      * For example see docs for {@link EclipseJdt}
      */
-    public JavaVersion getTargetCompatibility() {
-        return getTargetCompatibilityProperty().get();
-    }
-
-    /**
-     * The target JVM to generate {@code .class} files for.
-     * <p>
-     * For example see docs for {@link EclipseJdt}
-     *
-     * @since 9.0
-     */
-    @Incubating
-    abstract public Property<JavaVersion> getTargetCompatibilityProperty();
+    abstract public JavaVersion getTargetCompatibility();
 
     /**
      * Sets target compatibility.
@@ -136,34 +105,16 @@ public abstract class EclipseJdt {
         setTargetCompatibility((Object) targetCompatibility);
     }
 
-    public void setTargetCompatibility(Object targetCompatibility) {
-        JavaVersion version = JavaVersion.toVersion(targetCompatibility);
-        if (version != null) {
-            this.getTargetCompatibilityProperty().set(version);
-        }
-    }
+    public abstract void setTargetCompatibility(Object targetCompatibility);
 
     /**
      * The name of the Java Runtime to use.
      * <p>
      * For example see docs for {@link EclipseJdt}
      */
-    public String getJavaRuntimeName() {
-        return getJavaRuntimeNameProperty().get();
-    }
+    public abstract String getJavaRuntimeName();
 
-    public void setJavaRuntimeName(String javaRuntimeName) {
-        this.getJavaRuntimeNameProperty().set(javaRuntimeName);
-    }
-
-    /**
-     * The name of the Java Runtime to use.
-     *
-     * @return The name of the Java Runtime to use.
-     * @since 9.0
-     */
-    @Incubating
-    abstract public Property<String> getJavaRuntimeNameProperty();
+    public abstract void setJavaRuntimeName(String javaRuntimeName);
 
     /**
      * See {@link #file(Action) }
