+++ platforms/ide/ide/src/main/java/org/gradle/plugins/ide/eclipse/model/internal/EclipseJdtInternal.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2025 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.plugins.ide.eclipse.model.internal;
+
+import org.gradle.api.Incubating;
+import org.gradle.api.JavaVersion;
+import org.gradle.api.provider.Property;
+import org.gradle.plugins.ide.eclipse.model.EclipseJdt;
+
+public interface EclipseJdtInternal {
+    /**
+     * The source Java language level.
+     * <p>
+     * For example see docs for {@link EclipseJdt}
+     *
+     * @since 9.0
+     */
+    @Incubating
+    Property<JavaVersion> getSourceCompatibilityProperty();
+
+    /**
+     * The target JVM to generate {@code .class} files for.
+     * <p>
+     * For example see docs for {@link EclipseJdt}
+     *
+     * @since 9.0
+     */
+    @Incubating
+    Property<JavaVersion> getTargetCompatibilityProperty();
+
+    /**
+     * The name of the Java Runtime to use.
+     *
+     * @return The name of the Java Runtime to use.
+     * @since 9.0
+     */
+    @Incubating
+    Property<String> getJavaRuntimeNameProperty();
+}
